server/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Redis
redis>=5.0.1

# Auth
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2

# AI/LLM
google-generativeai>=0.3.2
langgraph>=0.0.24

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0

# Logging & Monitoring
loguru>=0.7.2
sentry-sdk>=1.39.1

# Utils
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0

# PDF Export (optional, pdflatex must be installed separately)
# For PDF generation, install TeX Live: brew install --cask mactex-no-gui
